requirements.txt
Flask==3.0.2
black==24.2.0
python-dotenv==1.0.1
Flask-SQLAlchemy==3.1.1
psycopg2-binary==2.9.9
alembic==1.13.1
Flask-Migrate==4.0.7